requirements.txt
snapcast
pycairo
PyGObject
python-dotenv
